% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bears}
\alias{bears}
\title{Bears phylogenetic tree}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
data(bears)
}
\description{
The example tree taken from the RevBayes tutorial website
}
\keyword{datasets}
