% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds.R
\name{birth_death_shift}
\alias{birth_death_shift}
\title{Title}
\usage{
birth_death_shift(phy, lambda, mu, eta, ntimes = 100)
}
\arguments{
\item{phy}{an object of class phylo}

\item{lambda}{vector of speciation rates}

\item{mu}{vector of extinction rates}

\item{eta}{the (common) rate of change between state i and state j. A scalar}

\item{ntimes}{number of solutions provided by the ODE solver for each branch}
}
\value{

}
\description{
Title
}
