% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_death_shift.R
\name{birth_death_shift2}
\alias{birth_death_shift2}
\title{Title}
\usage{
birth_death_shift2(phy, lambda, mu, eta, ntimes = 100)
}
\arguments{
\item{phy}{an object of class phylo}

\item{lambda}{vector of speciation rates}

\item{mu}{vector of extinction rates}

\item{eta}{the (common) rate of change between state i and state j. A scalar}

\item{ntimes}{number of solutions provided by the ODE solver for each branch}
}
\value{

}
\description{
Title
}
\examples{

data(bears)

lambda <- c(0.1, 0.2)
mu <- c(0.05, 0.15)

eta <- 0.05

res <- birth_death_shift2(bears, lambda, mu, eta)
}
