% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl_wrapper.R
\name{birth_death_shift3}
\alias{birth_death_shift3}
\title{Title}
\usage{
birth_death_shift3(phy, lambda, mu, eta, rho, verbose = FALSE)
}
\arguments{
\item{phy}{an object of class phylo}

\item{lambda}{vector of speciation rates}

\item{mu}{vector of extinction rates}

\item{eta}{the (common) rate of change between state i and state j. A scalar}
}
\value{

}
\description{
Title
}
\examples{
library(BDS)
library(ape)
library(ggtree)
library(tidytree)
library(ggplot2)
library(dplyr)

data(bears)

lambda <- c(0.1, 0.2)
mu <- c(0.05, 0.15)
rho <- 1.0
eta <- 0.05

res <- birth_death_shift3(bears, lambda, mu, eta, rho)
lambda_average <- res$lambda

th <- max(node.depth.edgelength(bears))
df1 <- tibble("node" = 1:max(bears$edge),
              "Speciation rate" = lambda_average)
x <- as_tibble(bears)

bearsdf <- merge(x, df1, by = "node")
td_bears <- as.treedata(bearsdf)

p1a <- ggtree(td_bears, aes(color = `Speciation rate`)) +
  geom_tiplab(size = 8) +
  theme(legend.position = c(0.2, 0.8)) +
  xlim(c(0.0, th + 10))
}
